refact: 숫자 최소/최대값 검증을 도메인 모델로 옮김
유효한 범위의 숫자값 검증
- 1 ~ 9 사이의 숫자 검증 로직을 도메인 모델(Ball)로 이동
