+++ src/main/java/baseball/model/Ball.java
@@ -3,10 +3,14 @@ package baseball.model;
 import java.util.Objects;
 
 public class Ball {
+	public static final int MIN_NUMBER = 1;
+	public static final int MAX_NUMBER = 9;
+
 	private int pitchingNumber;
 	private int position;
 
 	public Ball(int pitchingNumber, int position) {
+		valid(pitchingNumber);
 		this.pitchingNumber = pitchingNumber;
 		this.position = position;
 	}
@@ -25,6 +29,12 @@ public class Ball {
 
 	private boolean isEqualNumber(int number) {
 		return pitchingNumber == number;
+	}
+
+	public static void valid(int pitchingNumber) {
+		if(pitchingNumber < MIN_NUMBER  || MAX_NUMBER < pitchingNumber) {
+			throw new IllegalArgumentException("입력값이 올바르지 않습니다.");
+		}
 	}
 
 	@Override
