+++ src/main/java/baseball/model/BallNumbers.java
@@ -5,8 +5,6 @@ import java.util.List;
 import java.util.Set;
 
 public class BallNumbers {
-	public static final int MIN_NUMBER = 1;
-	public static final int MAX_NUMBER = 9;
 	public static final int DIGIT = 3;
 
 	public static void validCipher(List<Integer> balls) {
@@ -23,12 +21,6 @@ public class BallNumbers {
 		}
 
 		if(pitchingSet.size() != DIGIT) {
-			throw new IllegalArgumentException("입력값이 올바르지 않습니다.");
-		}
-	}
-
-	public static void validScale(int pitchingNumber) {
-		if(pitchingNumber < MIN_NUMBER  || MAX_NUMBER < pitchingNumber) {
 			throw new IllegalArgumentException("입력값이 올바르지 않습니다.");
 		}
 	}
